requirements.txt
numpy~=1.26.4
Flask~=3.0.2
tensorflow~=2.15.0
pillow~=10.2.0
opencv-python~=4.9.0.80
keras~=2.15.0
tqdm~=4.66.1
matplotlib~=3.8.2
lime~=0.2.0.1
scikit-image~=0.22.0